feat(xendit): handle billing plans

+++ includes/ajax/admin/pro.php
@@ -76,6 +76,28 @@ try {
       } else {
         $_POST['boost_pages'] = 0;
       }
+      /* Xendit billing plan */
+      $xendit_billing_plan = $package['xendit_billing_plan'];
+      $xendit_recurring_enabled = $system['xendit_enabled'];
+      if ($xendit_recurring_enabled && $_POST['price'] > 0 && $_POST['period'] != "life") {
+        /* check if Xendit billing plan is not created */
+        if (is_empty($xendit_billing_plan)) {
+          /* create Xendit billing plan */
+          $xendit_billing_plan = xendit_create_billing_plan($_POST['period_num'], strtoupper($_POST['period']), $_POST['price']);
+        } else {
+          /* check if the (period || period_num || price) is edited */
+          if ($package['period'] != $_POST['period'] || $package['period_num'] != $_POST['period_num'] || $package['price'] != $_POST['price']) {
+            /* replace the plan */
+            $xendit_billing_plan = xendit_replace_billing_plan($package['xendit_billing_plan'], $_POST['period_num'], strtoupper($_POST['period']), $_POST['price']);
+          }
+        }
+      } else {
+        $xendit_billing_plan = 'NULL';
+        if ($xendit_recurring_enabled && $package['xendit_billing_plan']) {
+          xendit_stop_billing_plan($package['xendit_billing_plan']);
+        }
+      }
+      $xendit_billing_plan = isset($xendit_billing_plan) ? $xendit_billing_plan : 'NULL';
       /* PayPal billing plan */
       $paypal_billing_plan = $package['paypal_billing_plan'];
       $paypal_recurring_enabled = $system['paypal_enabled'] && !is_empty($system['paypal_webhook']);
@@ -129,11 +151,11 @@ try {
         }
       }
       /* remove all users recurring payments */
-      if ($paypal_billing_plan == '' && $stripe_billing_plan == '') {
+      if ($paypal_billing_plan == '' && $stripe_billing_plan == '' && $xendit_billing_plan == '') {
         $db->query(sprintf("DELETE FROM users_recurring_payments WHERE handle = 'packages' AND handle_id = %s", secure($_GET['id'], 'int')));
       }
       /* update */
-      $db->query(sprintf("UPDATE packages SET name = %s, price = %s, period_num = %s, period = %s, color = %s, icon = %s, package_permissions_group_id = %s, allowed_videos_categories = %s, allowed_blogs_categories = %s, allowed_products = %s, verification_badge_enabled = %s, boost_posts_enabled = %s, boost_posts = %s, boost_pages_enabled = %s, boost_pages = %s, custom_description = %s, package_order = %s, paypal_billing_plan = %s, stripe_billing_plan = %s WHERE package_id = %s", secure($_POST['name']), secure($_POST['price']), secure($_POST['period_num']), secure($_POST['period']), secure($_POST['color']), secure($_POST['icon']), secure($_POST['permissions_group']), secure($_POST['allowed_videos_categories'], 'int'), secure($_POST['allowed_blogs_categories'], 'int'), secure($_POST['allowed_products'], 'int'), secure($_POST['verification_badge_enabled']), secure($_POST['boost_posts_enabled']), secure($_POST['boost_posts'], 'int'), secure($_POST['boost_pages_enabled']), secure($_POST['boost_pages'], 'int'), secure($_POST['custom_description']), secure($_POST['package_order'], 'int'), secure($paypal_billing_plan), secure($stripe_billing_plan), secure($_GET['id'], 'int')));
+      $db->query(sprintf("UPDATE packages SET name = %s, price = %s, period_num = %s, period = %s, color = %s, icon = %s, package_permissions_group_id = %s, allowed_videos_categories = %s, allowed_blogs_categories = %s, allowed_products = %s, verification_badge_enabled = %s, boost_posts_enabled = %s, boost_posts = %s, boost_pages_enabled = %s, boost_pages = %s, custom_description = %s, package_order = %s, paypal_billing_plan = %s, stripe_billing_plan = %s, xendit_billing_plan = %s WHERE package_id = %s", secure($_POST['name']), secure($_POST['price']), secure($_POST['period_num']), secure($_POST['period']), secure($_POST['color']), secure($_POST['icon']), secure($_POST['permissions_group']), secure($_POST['allowed_videos_categories'], 'int'), secure($_POST['allowed_blogs_categories'], 'int'), secure($_POST['allowed_products'], 'int'), secure($_POST['verification_badge_enabled']), secure($_POST['boost_posts_enabled']), secure($_POST['boost_posts'], 'int'), secure($_POST['boost_pages_enabled']), secure($_POST['boost_pages'], 'int'), secure($_POST['custom_description']), secure($_POST['package_order'], 'int'), secure($paypal_billing_plan), secure($stripe_billing_plan), secure($xendit_billing_plan), secure($_GET['id'], 'int')));
       /* return */
       return_json(['success' => true, 'message' => __("Package info have been updated")]);
       break;
@@ -170,6 +192,13 @@ try {
       } else {
         $_POST['boost_pages'] = 0;
       }
+      /* Xendit billing plan */
+      $xendit_billing_plan = 'NULL';
+      $xendit_recurring_enabled = $system['xendit_enabled'];
+      if ($xendit_recurring_enabled && $_POST['price'] > 0 && $_POST['period'] != "life") {
+        /* create Xendit billing plan */
+        $xendit_billing_plan = xendit_create_billing_plan($_POST['period_num'], strtoupper($_POST['period']), $_POST['price']);
+      }
       /* PayPal billing plan */
       $paypal_billing_plan = '';
       $paypal_recurring_enabled = $system['paypal_enabled'] && !is_empty($system['paypal_webhook']);
@@ -185,7 +214,7 @@ try {
         $stripe_billing_plan = stripe_create_billing_plan($_POST['name'], $_POST['custom_description'], $_POST['period_num'], $_POST['period'], $_POST['price']);
       }
       /* insert */
-      $db->query(sprintf("INSERT INTO packages (name, price, period_num, period, color, icon, package_permissions_group_id, allowed_videos_categories, allowed_blogs_categories, allowed_products, verification_badge_enabled, boost_posts_enabled, boost_posts, boost_pages_enabled, boost_pages, custom_description, package_order, paypal_billing_plan, stripe_billing_plan) VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)", secure($_POST['name']), secure($_POST['price']), secure($_POST['period_num']), secure($_POST['period']), secure($_POST['color']), secure($_POST['icon']), secure($_POST['permissions_group'], 'int'), secure($_POST['allowed_videos_categories'], 'int'), secure($_POST['allowed_blogs_categories'], 'int'), secure($_POST['allowed_products'], 'int'), secure($_POST['verification_badge_enabled']), secure($_POST['boost_posts_enabled']), secure($_POST['boost_posts'], 'int'), secure($_POST['boost_pages_enabled']), secure($_POST['boost_pages'], 'int'), secure($_POST['custom_description']), secure($_POST['package_order'], 'int'), secure($paypal_billing_plan), secure($stripe_billing_plan)));
+      $db->query(sprintf("INSERT INTO packages (name, price, period_num, period, color, icon, package_permissions_group_id, allowed_videos_categories, allowed_blogs_categories, allowed_products, verification_badge_enabled, boost_posts_enabled, boost_posts, boost_pages_enabled, boost_pages, custom_description, package_order, paypal_billing_plan, stripe_billing_plan, xendit_billing_plan) VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)", secure($_POST['name']), secure($_POST['price']), secure($_POST['period_num']), secure($_POST['period']), secure($_POST['color']), secure($_POST['icon']), secure($_POST['permissions_group'], 'int'), secure($_POST['allowed_videos_categories'], 'int'), secure($_POST['allowed_blogs_categories'], 'int'), secure($_POST['allowed_products'], 'int'), secure($_POST['verification_badge_enabled']), secure($_POST['boost_posts_enabled']), secure($_POST['boost_posts'], 'int'), secure($_POST['boost_pages_enabled']), secure($_POST['boost_pages'], 'int'), secure($_POST['custom_description']), secure($_POST['package_order'], 'int'), secure($paypal_billing_plan), secure($stripe_billing_plan), secure($xendit_billing_plan)));
       /* return */
       return_json(['callback' => 'window.location = "' . $system['system_url'] . '/' . $control_panel['url'] . '/pro/packages";']);
       break;
@@ -195,5 +224,8 @@ try {
       break;
   }
 } catch (Exception $e) {
-  return_json(['error' => true, 'message' => $e->getMessage()]);
+  return_json([
+    'error' => true,
+    'message' => $e->getMessage(),
+  ]);
 }
